Create README.md
As a graduate of Anadolu Technical High School Computer Department, I wanted to realize my dream of university and change the flow of my life within the conditions of the day. In this direction, despite the deduction of points due to the type of high school, I succeeded. During my education process, I developed myself in research, self-motivation, entrepreneurship and acquiring different abilities. I would like to state that I am extremely devoted and enthusiastic about being effective and fast in business life, and that I want to take on the responsibility of the position with curiosity and excitement. I am aware that willing, attentive and careful work will definitely bring success.